Update: plotroutines.py
This will not have much to do with a future generate_latex_pot.py. Anyway, I need to modify these files to have a better understanding of how to translate everything directly into LaTeX for the generator to be built in the future.

However, for now I will wait for Arancia's DOT files to have a better understanding and then proceed accordingly. For now, I will stick to working with the DOL files.